requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0
optuna>=3.0.0

# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0

# Model Management
mlflow>=2.0.0

# Database and Caching
redis>=4.0.0
psycopg2-binary>=2.9.0

# NLP and Text Processing
nltk>=3.8
textblob>=0.17.0
vaderSentiment>=3.3.0

# Geolocation and Mapping
geopy>=2.3.0
folium>=0.14.0

# Web and HTTP
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# Data Validation and Serialization
pydantic>=1.10.0
marshmallow>=3.19.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging and Monitoring
loguru>=0.6.0
prometheus-client>=0.16.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.5.0

# Environment Management
python-dotenv>=1.0.0
